Add distribution and win combination tests refactor a bit added bash script to run and build project

+++ src/main/java/com/cyberspeed/components/GameBoard.java
@@ -1,6 +1,8 @@
 package com.cyberspeed.components;
 
+import java.math.BigDecimal;
+
 public interface GameBoard<R> {
 
-    R bet(double bettingAmount);
+    R bet(BigDecimal bettingAmount);
 }
