+++ src/main/java/com/cyberspeed/components/RewardCalculator.java
@@ -2,7 +2,9 @@ package com.cyberspeed.components;
 
 import com.cyberspeed.entities.ScratchGameResult;
 
+import java.math.BigDecimal;
+
 public interface RewardCalculator {
 
-    ScratchGameResult calculate(String[][] board, double bettingAmount);
+    ScratchGameResult calculate(String[][] board, BigDecimal bettingAmount);
 }
