+++ src/main/java/com/cyberspeed/entities/ScratchGameResult.java
@@ -1,11 +1,12 @@
 package com.cyberspeed.entities;
 
+import java.math.BigDecimal;
 import java.util.List;
 import java.util.Map;
 
 public record ScratchGameResult(
         String[][] matrix,
-        double reward,
+        BigDecimal reward,
         Map<String, List<String>> appliedWinningCombinations,
         String appliedBonusSymbol
 ) {
