+++ src/main/java/com/cyberspeed/utils/InputParameters.java
@@ -1,9 +1,10 @@
 package com.cyberspeed.utils;
 
 import java.io.File;
+import java.math.BigDecimal;
 
 public record InputParameters(
         File config,
-        Double bettingAmount
+        BigDecimal bettingAmount
 ) {
 }
